backend/requirements.txt
# ── Core Web Framework ─────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
loguru==0.7.2
httpx==0.27.0
python-dotenv==1.0.1
python-dateutil==2.9.0
google-generativeai>=0.8.0

# ── Database ───────────────────────────────────────────────────────────────
motor==3.4.0

# ── Document Processing ────────────────────────────────────────────────────
PyMuPDF==1.24.5
pdfplumber==0.11.1
python-docx==1.1.2
reportlab==4.2.2
openpyxl==3.1.5
pandas==2.2.2
Pillow==10.4.0

# ── ML Core ───────────────────────────────────────────────────────────────
numpy==1.26.4
scikit-learn==1.5.1
xgboost==2.1.0
lightgbm==4.4.0
catboost==1.2.5
shap==0.45.1
joblib==1.4.2
scipy==1.14.0

# ── Imbalanced Learning ───────────────────────────────────────────────────
imbalanced-learn==0.12.3

# ── Hyperparameter Tuning ─────────────────────────────────────────────────
optuna==3.6.1

# ── Graph Intelligence ────────────────────────────────────────────────────
networkx==3.3

# ── Web Scraping ──────────────────────────────────────────────────────────
beautifulsoup4==4.12.3
lxml==5.2.2

# ── OCR (heavy — install separately if needed) ────────────────────────────
# paddlepaddle==2.6.1
# paddleocr==2.7.3
# playwright==1.45.0

# ── NLP (heavy — install separately) ─────────────────────────────────────
# spacy==3.7.5
# transformers==4.42.4
# torch==2.3.1
